Refactor

+++ css-modules-automation/codemod/src/utils/renameStylesheetFile.ts
@@ -1,9 +1,13 @@
-import { renameSync } from "fs";
+import { existsSync, renameSync } from "fs";
+import { printWithCondition } from "./print";
 
 export const getCssModulesFileName = (path: string) => {
   return path.replace(".scss", ".module.scss");
 };
 
 export const renameStylesheetFile = (filename: string) => {
-  renameSync(filename, getCssModulesFileName(filename));
+  if (existsSync(filename)) {
+    renameSync(filename, getCssModulesFileName(filename));
+    printWithCondition(true, `^^^ renameStylesheetFile, ${filename} -> ${getCssModulesFileName(filename)}`);
+  }
 };
